requirements.txt
# Python Dependencies for docuMINE
# =================================
# These are used by the AI quoting automation features.
# Node.js dependencies are in package.json (main application)

# Browser Use - AI-powered browser automation (Story Q7.1)
browser-use>=0.11.0
langchain-anthropic>=1.3.0

# Playwright - Browser automation (required by browser-use)
playwright>=1.40.0

# Environment management
python-dotenv>=1.0.0
